src/controllers/solana-controller.ts
import { Request, Response } from 'express';

class SolanaController {
  // Route to fetch data from Solana
  async fetchData(req: Request, res: Response) {
    try {
      const { accountAddress } = req.params;

      // Replace this with your Solana data fetching logic
      const data = await fetchDataFromSolana(accountAddress);

      res.status(200).json({ data });
    } catch (error) {
      console.error('Error fetching data from Solana:', error);
      res.status(500).json({ error: 'Internal Server Error' });
    }
  }

  // Route to send a transaction to Solana
  async sendTransaction(req: Request, res: Response) {
    try {
      const { privateKey, instructions } = req.body;

      // Replace this with your Solana transaction sending logic
      const transactionHash = await sendTransactionToSolana(privateKey, instructions);

      res.status(200).json({ transactionHash });
    } catch (error) {
      console.error('Error sending transaction to Solana:', error);
      res.status(500).json({ error: 'Internal Server Error' });
    }
  }
}

export default SolanaController;
