src/tests/solana-controller.test.ts
import { Request, Response } from 'express';
import SolanaController from './solana-controller'; // Import your Solana controller here

// Mocked Express Request and Response objects
const mockedRequest = {} as Request;
const mockedResponse = {} as Response;

// Mocked Express `next` function
const next = jest.fn();

// Initialize your Solana controller
const solanaController = new SolanaController();

describe('Solana Controller Tests', () => {
  beforeAll(() => {
    // Setup any necessary configurations or test data
  });

  afterAll(() => {
    // Cleanup resources if needed
  });

  test('Fetch Data from Solana', async () => {
    // Mock any necessary request parameters or body
    mockedRequest.params = {
      accountAddress: '5P3giWpPBrVKL8QP8roKM7NsLdi3ie1Nc2b5r9mGtvwb', // Replace with a valid account address
    };

    // Mock any expected behavior or responses from your controller
    const sendMock = jest.fn();
    mockedResponse.status = jest.fn().mockReturnValue({ send: sendMock });

    // Call your Solana controller method for fetching data
    await solanaController.fetchData(mockedRequest, mockedResponse, next);

    // Write your assertions to test the response or side effects
    expect(sendMock).toHaveBeenCalled();
    // Add more assertions based on your specific expectations
  });

  test('Send Transaction to Solana', async () => {
    // Mock any necessary request parameters or body
    mockedRequest.body = {
      privateKey: 'your_private_key_here', // Replace with a valid private key
      instructions: [
        // Define your transaction instructions here
        // Example: Create, update, or interact with a Solana program
      ],
    };

    // Mock any expected behavior or responses from your controller
    const sendMock = jest.fn();
    mockedResponse.status = jest.fn().mockReturnValue({ send: sendMock });

    // Call your Solana controller method for sending a transaction
    await solanaController.sendTransaction(mockedRequest, mockedResponse, next);

    // Write your assertions to test the response or side effects
    expect(sendMock).toHaveBeenCalled();
    // Add more assertions based on your specific expectations
  });
});
